feat(schemas): Add 'list' schema, improve validation logic, improve map validation
Maps and lists can now have validators for the main data, not only for sub-data. For example, a list can have a `unique` boolean validator that checks the lists items for uniqueness.

`validate.scss` has been cleaned up, so that there is now a `_run-validators` method that should work for all validators equally.

`base.validate` is now called in the root `_validate` function, which will immediately return if there are any type errors for validators. This means that afterwards, the validate functions don't have to care about that anymore.

Also extended unit tests